Huge Update
This update involved a lot of work behind the scenes to get Head's method working, then get Thwaites working with the same solver, then abstracting out the similarities into a parent class. The SimData class now creates a cubic spline of the velocity distribution data when initialized. As the RK45 solver steps through for each method, it can call this spline to obtain a current estimate of the velocity at a given x instead of getting it from a point. Using RK for Thwaites is largely unnecessary because it's just an integral and  not a differential equation. In the derivatives function for thwaites, y isn't even used.  The user now obtains values of interest at any point using functions of x. The user calls HeadSim.step() (for example) to step the simulation forward, and can see the status of the simulation (direct from RK45().status()) to see how the simulation is doing.