Updated flatplate_mild_pressure_gradient.py - Plots
The updated plotting sections use matplotlib.pyplot's mathtext functionality to make more porfessional plots. It should be noted that the way this is set up changes the user's default plotting options, so something should be added to revert them...

In addition, mathtext's abilities seem to be limited so that italics are only supported in sans-serif font. This is why the variables do not appear italicized in the figures. There is an option to override mathtext and use actual LaTeX, which may involve the user running actual LaTeX on their computer. For the time being, I haven't delved deep enough into this functionality, but for the actual thesis writeup, this will be implemented for consistency, unless mathtext has this function and  I ahven't found it.