Added all figure files
all_figures.py is intended to produce all figures but does not run the XFOIL cases yet. thesis_classes.vpp is a Visual Paradigm file that contains the two class diagrams in the thesis. analytical_thwaites_wedge.py contains a function that implements Falkner-Skan flow given a value of m and other freestream conditions. The XFOIL files need XFOIL to be installed in the same directory as pyBL. There should also be a folder called "figures" for plot_BL_params.py to put all of the figures in. plot_BL_params.py is a function used to make 4-subplot figures showing momentum thickness, displacement thickness, skin friction coefficient, and shape factor with a legend describing which lines in each plot are what simulation.